Add GPU support
AT
This program can now run on the GPU through CUDA

Also, the program will re-dump the initial condition with updated accelerations and internal energy to a new HDF5 file, which can then be used as input for future simulations or diagnostics.

Besides, the timer is now using nanosecond for timing since GPU is so fast that the microsecond cannot handle anymore.

Both Windows and Linux/macOS Makefile (for windows: Makefile.py) are updated for GPU support

Also, the `README.md` has updated